% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.drawEnhancedMeanPlot}
\alias{.drawEnhancedMeanPlot}
\title{Internal plotting function}
\usage{
.drawEnhancedMeanPlot(y, g, ylab = "Response", xlab = "Group")
}
\description{
Internal plotting function
}
