% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{meanPlotEnhanced}
\alias{meanPlotEnhanced}
\title{Enhanced mean plot using the active R Commander dataset}
\usage{
meanPlotEnhanced()
}
\description{
Enhanced mean plot using the active R Commander dataset
}
